email: fix save attachment to storage for pop3 accounts
AttachmentUtilities#saveAttachment is never called for pop3 accounts, so the
attachment are never saved to storage when is requested by the user.
This change call saveAttachment method when the user request a sync of the pop3 account
and the document was saved previously in the internal database storage.

Change-Id: If5a857eb40b7153ef9f1278e78b98f63e35b431d
JIRA: CYAN-3762
Issue: https://jira.cyanogenmod.org/browse/CYAN-3762
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>